Create Problem26
Grandfather's yard has a rectangular shape, and its length and width are known. Create a function labeled bunicul that has 2 parameters representing the length and width of the yard and displays the area of this yard, as well as the length of the fence surrounding the yard (these will be separated by a new line).