Create Problem38
Consider a word formed from at least two and at most 100 characters, only lowercase letters of the English alphabet. Write a program that reads from the keyboard a word of the specified type and displays on the screen the message YES if the word contains only consonants and, possibly, the vowel i, or the message NO otherwise.