Create Problem43
Read a string of characters representing a sentence that contains uppercase and lowercase letters of the English alphabet, spaces, digits, and other symbols. In the following, we consider a word any sequence delimited by spaces that contains at least one letter. Display the read string so that the words of maximum length are replaced with their reverse (mirror image), the rest of the words and their arrangement in the sentence to remain unchanged.