Create Problem62
A word is a palindrome if it reads the same in both directions. For example, "racecar" is a palindrome. A sentence is palindromic if, by eliminating spaces and other punctuation marks, it becomes a palindrome. For example, "A man, a plan, a canal, Panama" is palindromic. Write a program that reads several sentences and determines about each if it is palindromic.