Create Problem101
If the students of a class sit n in a bench, e students will stand up, and if they sit m in a bench, b benches will be unoccupied. What is the number of students and benches in the class, respectively? The program will display on the first line the number of students and on the next line the number of benches in the class.