Create Problem106
Delete String Sequences / BirdLanguage1
A sentence is written in bird language. Translate it into Romanian. The program reads a character string s from the keyboard. The program will display the string t on the screen, representing the translation of the string s into Romanian. The strings s and t will have at most 255 characters, the string s will contain lowercase letters of the English alphabet and spaces, the vowels are: aeiou, and the string given in bird language must be correct