Create Problem108
 CapsLock
Marcel got a new keyboard for his birthday. He wants to write a text using only letters of the English alphabet and spaces. He discovered the CapsLock key, which when pressed is activated. If the key is activated, then all the letters he will write afterwards, until the key is deactivated (by pressing CapsLock again) will be uppercase.