Create Problem124
 Largest Remainder
For a natural number m, the largest remainder is the greatest remainder obtained by dividing the number m by all natural numbers from 1 to m. Given a natural number n, determine for each number from 1 to n the largest remainder, and these largest remainders are summed up. The task is to find this sum.